+++ yash-syntax/src/parser/lex.rs
@@ -40,6 +40,7 @@ mod word;
 pub use self::braced_param::is_name_char;
 pub use self::core::*;
 pub use self::keyword::Keyword;
+pub use self::keyword::ParseKeywordError;
 pub use self::op::is_operator_char;
 pub use self::op::Operator;
 pub use self::raw_param::is_portable_name_char;